restructured routing to better differentiate between dashboard and landing views and fix nav issues
