+++ src/views/landing/Tenant.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import WorkorderForm from '../components/WorkorderForm/index';
+import WorkorderForm from '../../components/WorkorderForm/index';
 
 const Tenant = () => {
   return (
