+++ src/views/landing/ThankYou.js
